serch feature added and merged
