+++ android/app/src/main/AndroidManifest.xml
@@ -1,9 +1,12 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android">
     <!-- access storage -->
-    <uses-permission android:name ="android-permission.READ_EXTERNAL_STORAGE"/>
-    <uses-permission android:name = "android-permission.WRITE_EXTERNAL_STORAGE"/>
+<!--    <uses-permission android:name ="android-permission.READ_EXTERNAL_STORAGE"/>-->
+<!--    <uses-permission android:name = "android-permission.WRITE_EXTERNAL_STORAGE"/>-->
+    <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE"/>
+    <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE"/>
+
     <application
-        android:label="msss"
+        android:label="Melone"
         android:name="${applicationName}"
         android:icon="@mipmap/ic_launcher">
         <activity
