+++ lib/features/authentication/login/login_page.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:Melone/features/local_music_player/local_music_player_page.dart';
 import '../../music_feed/music_feed.dart';
-import '../../music_feed/home_feed.dart';
+import '../../music_feed/music_feed_page.dart';
 
 class LoginPage extends StatefulWidget {
   @override
