+++ pubspec.yaml
@@ -42,7 +42,7 @@ dependencies:
   permission_handler: ^11.3.1
   get: ^4.6.6
   http: ^1.2.1
-  # permission_handler: ^8.1.4+2
+
 
 dev_dependencies:
   flutter_test:
